+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Next release
 * Feat: [Android] Turn off the screen when a call is active and the head is against the handset. @solid-software (https://solid.software)
 * UPDATE: Hold call signature changed to `setHoldCall({bool = true})`, with query check `isHolding()`
+* ADD: Added additional catches for missed / declined calls/
 
 ## 0.0.9
 * Feat: forwarded callInvite custom parameters to flutter